refactor: extract RepayDialog logic into useRepayLease (#185)
RepayDialog.vue held ~520 script LOC (asset/repayment derivation,
balance validation, the repay submit path, two watchers) in one SFC,
against the house rule that a template be a thin projection of
already-computed values.

Move that logic verbatim into a co-located useRepayLease() composable
and the preview block into a props-only RepayPreview.vue; RepayDialog.vue
becomes template plus composable wiring. Behaviour is unchanged and the
existing RepayDialog.test.ts (11 cases) passes untouched.

The inline close handler and the two slider-label arrows become named
composable handlers (closeDialog, onSelectFullDebt, onSelectZero) so the
template carries no logic. closeDialog uses === where the template used
== (both operands are strings, identical result) to satisfy eqeqeq,
which does not apply inside .vue template expressions.

Continues the CloseDialog split; same pattern.
