+++ src/modules/leases/components/single-lease/useRepayLease.ts
@@ -1,0 +1,561 @@
+import { computed, inject, onBeforeUnmount, onMounted, ref, watch } from "vue";
+import { useRoute, useRouter } from "vue-router";
+import type { Dialog } from "web-components";
+import { ToastType } from "web-components";
+import { RouteNames } from "@/router";
+
+import { useWalletStore } from "@/common/stores/wallet";
+import { useBalancesStore } from "@/common/stores/balances";
+import { useConfigStore } from "@/common/stores/config";
+import { usePricesStore } from "@/common/stores/prices";
+import { useHistoryStore } from "@/common/stores/history";
+import { useLeasesStore, type LeaseDisplayData } from "@/common/stores/leases";
+import { classifyError, getMicroAmount, LeaseMath, Logger, walletOperation } from "@/common/utils";
+import { formatNumber, formatDecAsUsd, formatUsd, formatTokenBalance } from "@/common/utils/NumberFormatUtils";
+import { getLpnByProtocol } from "@/common/utils/CurrencyLookup";
+import { NATIVE_CURRENCY } from "../../../../config/global/network";
+import type { ExternalCurrency } from "@/common/types";
+import { minimumLeaseAmount } from "@/config/global";
+import type { AssetBalance } from "@/common/stores/wallet/types";
+import { CoinPretty, Dec, Int } from "@keplr-wallet/unit";
+
+import type { NolusWallet } from "@nolus/nolusjs";
+import { CurrencyUtils, NolusClient } from "@nolus/nolusjs";
+import { useI18n } from "vue-i18n";
+import type { Coin } from "@cosmjs/proto-signing";
+import { Lease } from "@nolus/nolusjs/build/contracts";
+import type { LeaseInfo } from "@/common/api";
+
+export function useRepayLease() {
+  const route = useRoute();
+  const router = useRouter();
+  const pricesStore = usePricesStore();
+  const walletStore = useWalletStore();
+  const balancesStore = useBalancesStore();
+  const configStore = useConfigStore();
+  const historyStore = useHistoryStore();
+  const leasesStore = useLeasesStore();
+  const i18n = useI18n();
+
+  const amount = ref("");
+  const amountErrorMsg = ref("");
+  const selectedCurrency = ref(0);
+  const isLoading = ref(false);
+  const swapFee = ref(0);
+  const sliderValue = ref(0);
+  const loading = ref(false);
+  const disabled = ref(false);
+  const sliderValueDec = ref(new Dec(0));
+
+  const reload = inject("reload", () => {});
+  const onShowToast = inject("onShowToast", (_data: { type: ToastType; message: string }) => {});
+
+  const dialog = ref<typeof Dialog | null>(null);
+  const lease = ref<LeaseInfo | null>(null);
+  const displayData = ref<LeaseDisplayData | null>(null);
+
+  function initLease() {
+    // Read from store cache — the parent already fetched this lease.
+    // Avoid calling fetchLeaseDetails here: it mutates store state which
+    // triggers the parent's watcher, re-renders, and unmounts this dialog.
+    const cached = leasesStore.getLease(route.params.id as string);
+    if (cached) {
+      lease.value = cached;
+      displayData.value = leasesStore.getLeaseDisplayData(cached);
+      if (cached.status === "closed") {
+        router.push(`/${RouteNames.LEASES}`);
+      }
+    }
+  }
+
+  watch(
+    () => configStore.initialized,
+    () => {
+      if (configStore.initialized) {
+        initLease();
+      }
+    },
+    { immediate: true }
+  );
+
+  onMounted(() => {
+    dialog?.value?.show();
+  });
+
+  onBeforeUnmount(() => {
+    dialog?.value?.close();
+  });
+  const currency = computed(() => {
+    return assets.value[selectedCurrency.value];
+  });
+
+  function onSetAmount(percent: number) {
+    const a = getRepayment(percent);
+    sliderValue.value = percent;
+    sliderValueDec.value = new Dec(percent);
+    amount.value = a?.repayment ? a.repayment.toString(currency.value.decimal_digits) : "";
+  }
+
+  function onSelectFullDebt() {
+    onSetAmount(100);
+  }
+
+  function onSelectZero() {
+    onSetAmount(0);
+  }
+
+  const assets = computed(() => {
+    const data = [];
+    for (const asset of (balances.value as ExternalCurrency[]) ?? []) {
+      const value = new Dec(asset.balance?.amount.toString() ?? 0, asset.decimal_digits);
+      const balance = formatTokenBalance(value);
+      const exactBalance = value.isZero() ? "0" : value.toString(asset.decimal_digits).replace(/\.?0+$/, "");
+      const denom = (asset as ExternalCurrency).ibcData ?? (asset as AssetBalance).from;
+      const price = new Dec(pricesStore.prices[asset.key]?.price ?? 0);
+      const stable = price.mul(value);
+
+      data.push({
+        name: asset.name,
+        value: denom,
+        label: asset.shortName,
+        shortName: asset.shortName,
+        icon: asset.icon,
+        decimal_digits: asset.decimal_digits,
+        balance: {
+          value: exactBalance,
+          customLabel: `${balance} ${asset.shortName}`,
+          ticker: asset.shortName,
+          denom: asset.balance.denom,
+          amount: asset.balance?.amount
+        },
+        ibcData: (asset as ExternalCurrency).ibcData,
+        native: asset.native,
+        symbol: asset.symbol,
+        ticker: asset.ticker,
+        key: asset.key,
+        stable,
+        price: formatDecAsUsd(stable)
+      });
+    }
+    return data;
+  });
+
+  const calculatedBalance = computed(() => {
+    const asset = assets.value[selectedCurrency.value];
+    if (!asset) {
+      return formatUsd(0);
+    }
+    const price = new Dec(pricesStore.prices[asset.key]?.price ?? 0);
+    const v = amount?.value?.length ? amount?.value : "0";
+    const stable = price.mul(new Dec(v));
+    return formatDecAsUsd(stable);
+  });
+
+  const balances = computed(() => {
+    if (!lease.value || lease.value.status !== "opened") return [];
+
+    const positionType = configStore.getPositionType(lease.value.protocol);
+    let repaymentCurrency;
+
+    if (positionType === "Short") {
+      // Short: debt is in LPN (e.g. USDC_NOBLE), use debt.ticker
+      const debtTicker = lease.value.debt.ticker;
+      repaymentCurrency = configStore.currenciesData[`${debtTicker}@${lease.value.protocol}`];
+    } else {
+      // Long: debt is in LPN (e.g. USDC)
+      repaymentCurrency = getLpnByProtocol(lease.value.protocol);
+    }
+
+    if (!repaymentCurrency) return [];
+    const resolvedCurrency = repaymentCurrency;
+
+    // Use the protocol-specific currency directly (not a cross-protocol find by ibcData)
+    // to ensure the .key matches the price entry in pricesStore.
+    const walletBalance = balancesStore.balances.find((item) => item.denom === resolvedCurrency.ibcData);
+    return [{ ...resolvedCurrency, balance: walletBalance ?? { denom: resolvedCurrency.ibcData, amount: "0" } }];
+  });
+
+  function handleAmountChange(event: string) {
+    amount.value = event;
+    if (amount.value !== "" && debt.value) {
+      let percent = new Dec(amount.value).quo(debt.value.amount.toDec()).mul(new Dec(100));
+      if (percent.isNegative()) {
+        percent = new Dec(0);
+      }
+      if (percent.gt(new Dec(100))) {
+        percent = new Dec(100);
+      }
+      sliderValueDec.value = percent;
+
+      sliderValue.value = Number(percent.toString(0));
+    }
+  }
+
+  const debt = computed(() => {
+    const selectedCurrency = currency.value;
+    if (selectedCurrency) {
+      const repaymentData = getRepayment(100);
+      if (!repaymentData) {
+        return undefined;
+      }
+      const { repayment } = repaymentData;
+      const repaymentInt = repayment.mul(new Dec(10).pow(new Int(selectedCurrency.decimal_digits))).truncate();
+
+      return {
+        amount: new CoinPretty(
+          {
+            coinDenom: selectedCurrency.shortName,
+            coinMinimalDenom: selectedCurrency.ibcData,
+            coinDecimals: selectedCurrency.decimal_digits
+          },
+          repaymentInt
+        ).hideDenom(true)
+      };
+    }
+    return undefined;
+  });
+
+  function getRepayment(p: number) {
+    if (!lease.value || lease.value.status !== "opened") return undefined;
+
+    const amount = outStandingDebt();
+    const lpn = getLpnByProtocol(lease.value.protocol);
+    if (!lpn) return undefined;
+    const c = lpn;
+
+    const amountToRepay = CurrencyUtils.convertMinimalDenomToDenom(
+      amount.toString(),
+      c.shortName,
+      c.ibcData,
+      c.decimal_digits
+    ).toDec();
+
+    const percent = new Dec(p).quo(new Dec(100));
+    let repaymentInStable = amountToRepay.mul(percent);
+    const selectedCurrency = currency.value;
+    if (!selectedCurrency) return undefined;
+
+    if (swapFee.value) {
+      repaymentInStable = repaymentInStable.add(repaymentInStable.mul(new Dec(swapFee.value)));
+    }
+
+    const positionType = configStore.getPositionType(lease.value.protocol);
+
+    if (positionType === "Short") {
+      const lpnShort = getLpnByProtocol(lease.value.protocol);
+      if (!lpnShort) return undefined;
+      const lpnPriceEntry = pricesStore.prices[lpnShort.key as string];
+      const selectedPriceEntry = pricesStore.prices[selectedCurrency.key as string];
+      // Missing price entries make the conversion undefined (and would divide by
+      // zero); bail so callers render an empty preview instead of throwing.
+      if (!lpnPriceEntry?.price || !selectedPriceEntry?.price) return undefined;
+      const price = new Dec(lpnPriceEntry.price);
+      const selected_asset_price = new Dec(selectedPriceEntry.price);
+      const repayment = repaymentInStable.mul(price);
+      return {
+        repayment: repayment.quo(selected_asset_price),
+        repaymentInStable: repayment,
+        selectedCurrencyInfo: selectedCurrency
+      };
+    } else {
+      const price = new Dec(pricesStore.prices[selectedCurrency.key as string]?.price ?? 1);
+      const repayment = repaymentInStable.quo(price);
+      return {
+        repayment,
+        repaymentInStable,
+        selectedCurrencyInfo: selectedCurrency
+      };
+    }
+  }
+
+  function outStandingDebt() {
+    if (!lease.value || lease.value.status !== "opened") return new Dec(0);
+
+    const debt = new Dec(lease.value.debt.principal)
+      .add(new Dec(lease.value.debt.overdue_margin))
+      .add(new Dec(lease.value.debt.overdue_interest))
+      .add(new Dec(lease.value.debt.due_margin))
+      .add(new Dec(lease.value.debt.due_interest));
+
+    return debt;
+  }
+
+  function isAmountValid() {
+    let isValid = true;
+    amountErrorMsg.value = "";
+
+    const amnt = amount.value;
+    const coinData = currency.value;
+    const leaseValue = lease.value;
+
+    if (coinData && leaseValue) {
+      if (amnt || amnt !== "") {
+        const priceEntry = pricesStore.prices[coinData.key as string];
+        // The price feed can lag the form (late WS price, config skew). Without it
+        // validation cannot run; surface a specific message rather than dereferencing
+        // a missing entry and throwing inside the reactive watch.
+        if (!priceEntry?.price) {
+          amountErrorMsg.value = i18n.t("message.unexpected-error");
+          return false;
+        }
+
+        const amountInMinimalDenom = CurrencyUtils.convertDenomToMinimalDenom(amnt, "", coinData.decimal_digits);
+        const balance = CurrencyUtils.calculateBalance(
+          priceEntry.price,
+          amountInMinimalDenom,
+          coinData.decimal_digits
+        ).toDec();
+        const minAmount = new Dec(minimumLeaseAmount);
+        const p = new Dec(priceEntry.price);
+        const amountInStable = new Dec(amnt.length === 0 ? "0" : amnt).mul(p);
+
+        const debt = getDebtValue();
+        const lpn = getLpnByProtocol(leaseValue.protocol);
+        const debtInCurrencies = debt.quo(new Dec(priceEntry.price));
+        const minAmm = new Dec(minimumLeaseAmount).mul(new Dec(10 ** lpn.decimal_digits));
+
+        const minAmountCurrency = minAmount.quo(p);
+        const isLowerThanOrEqualsToZero = new Dec(amountInMinimalDenom.amount || "0").lte(new Dec(0));
+
+        const isGreaterThanWalletBalance = new Int(amountInMinimalDenom.amount.toString() || "0").gt(
+          new Int(coinData?.balance?.amount ?? "0")
+        );
+
+        if (isLowerThanOrEqualsToZero) {
+          amountErrorMsg.value = i18n.t("message.invalid-balance-low");
+          isValid = false;
+        }
+
+        if (isGreaterThanWalletBalance) {
+          amountErrorMsg.value = i18n.t("message.invalid-balance-big");
+          isValid = false;
+        }
+
+        if (debt.gt(minAmm) && amountInStable.lt(minAmount)) {
+          amountErrorMsg.value = i18n.t("message.min-amount-allowed", {
+            amount: formatTokenBalance(minAmountCurrency),
+            currency: coinData.shortName
+          });
+          isValid = false;
+        }
+
+        const b = CurrencyUtils.convertDenomToMinimalDenom(
+          balance.toString(),
+          "",
+          coinData.decimal_digits
+        ).amount.toDec();
+        if (b.gt(debt) && debt.gt(minAmountCurrency)) {
+          const n = new Dec(debtInCurrencies.truncate().toString(), coinData.decimal_digits);
+          amountErrorMsg.value = i18n.t("message.lease-only-max-error", {
+            maxAmount: formatTokenBalance(n),
+            symbol: coinData.shortName
+          });
+          isValid = false;
+        }
+      } else {
+        amountErrorMsg.value = i18n.t("message.missing-amount");
+        isValid = false;
+      }
+    }
+
+    return isValid;
+  }
+
+  function getDebtValue() {
+    let debt = outStandingDebt();
+    debt = debt.add(debt.mul(new Dec(swapFee.value)));
+
+    if (!lease.value) {
+      throw new Error("Lease not available");
+    }
+    const positionType = configStore.getPositionType(lease.value.protocol);
+    if (positionType === "Short") {
+      const lpn = getLpnByProtocol(lease.value.protocol);
+      if (!lpn) {
+        throw new Error("LPN currency not available");
+      }
+      const priceEntry = pricesStore.prices[lpn.key as string];
+      if (!priceEntry?.price) {
+        throw new Error("LPN price not available");
+      }
+      return debt.mul(new Dec(priceEntry.price));
+    }
+    return debt;
+  }
+
+  async function onSendClick() {
+    try {
+      disabled.value = true;
+      await walletOperation(repayLease);
+    } catch (e: unknown) {
+      Logger.error(e);
+      amountErrorMsg.value = i18n.t(classifyError(e));
+    } finally {
+      disabled.value = false;
+    }
+  }
+
+  async function repayLease() {
+    const wallet = walletStore.wallet as NolusWallet;
+    if (wallet && isAmountValid() && lease.value) {
+      try {
+        loading.value = true;
+
+        const microAmount = getMicroAmount(currency.value.balance.denom, amount.value);
+        const funds: Coin[] = [
+          {
+            denom: microAmount.coinMinimalDenom,
+            amount: microAmount.mAmount.amount.toString()
+          }
+        ];
+
+        const cosmWasmClient = await NolusClient.getInstance().getCosmWasmClient();
+        const leaseClient = new Lease(cosmWasmClient, lease.value.address);
+
+        const { txBytes } = await leaseClient.simulateRepayLeaseTx(wallet, funds);
+        await walletStore.wallet?.broadcastTx(txBytes as Uint8Array);
+        leasesStore.markLeaseInProgress(lease.value.address, "repayment");
+        balancesStore.fetchBalances();
+        historyStore.loadActivities();
+        reload();
+        dialog?.value?.close();
+        onShowToast({
+          type: ToastType.success,
+          message: i18n.t("message.toast-repaid")
+        });
+      } catch (e: unknown) {
+        amountErrorMsg.value = i18n.t(classifyError(e));
+        Logger.error(e);
+      } finally {
+        loading.value = false;
+      }
+    }
+  }
+
+  // Reacts to: the typed amount and the selected repayment currency.
+  // Re-firing on the same input is safe — isAmountValid is idempotent (it resets
+  // amountErrorMsg on each run). The try/catch is the backstop for any residual
+  // throw from a lagging price/currency feed so the field never silently freezes.
+  watch(
+    () => [amount.value, selectedCurrency.value],
+    () => {
+      try {
+        isAmountValid();
+      } catch (e) {
+        Logger.error(e);
+        amountErrorMsg.value = i18n.t("message.unexpected-error");
+      }
+    }
+  );
+
+  function getPrice() {
+    if (!lease.value || !displayData.value) return new Dec(0);
+    return displayData.value.openingPrice;
+  }
+
+  const detbPartial = computed(() => {
+    const price = getPrice();
+    const debt = getRepayment(Number(sliderValueDec.value.toString()));
+    const debtTotal = getRepayment(100);
+
+    const d = debt?.repayment;
+    if (price && d && debtTotal && lease.value) {
+      const ticker = lease.value.etl_data?.lease_position_ticker ?? lease.value.amount.ticker;
+      const currecy = configStore.currenciesData[`${ticker}@${lease.value.protocol}`];
+
+      const positionType = configStore.getPositionType(lease.value.protocol);
+      const rest = debtTotal.repayment.sub(d);
+
+      if (positionType === "Short") {
+        return {
+          payment: `${formatTokenBalance(d)} ${currecy?.shortName ?? ""}`,
+          rest: `${formatTokenBalance(rest)} ${currecy?.shortName ?? ""}`
+        };
+      } else {
+        const lpn = getLpnByProtocol(lease.value.protocol);
+        return {
+          payment: `${formatTokenBalance(d)} ${lpn.shortName}`,
+          rest: `${formatTokenBalance(rest)} ${lpn.shortName}`
+        };
+      }
+    }
+
+    return { payment: "", rest: "" };
+  });
+
+  const debtData = computed(() => {
+    const price = getPrice();
+    const debt = getRepayment(100);
+    const d = debt?.repayment;
+
+    if (price && d && lease.value) {
+      const ticker = lease.value.etl_data?.lease_position_ticker ?? lease.value.amount.ticker;
+      const currecy = configStore.currenciesData[`${ticker}@${lease.value.protocol}`];
+
+      const positionType = configStore.getPositionType(lease.value.protocol);
+      if (positionType === "Short") {
+        return `${formatTokenBalance(d)} ${currecy?.shortName ?? ""}`;
+      } else {
+        const lpn = getLpnByProtocol(lease.value.protocol);
+        return `${formatTokenBalance(d)} ${lpn.shortName}`;
+      }
+    }
+
+    return "";
+  });
+
+  const liquidation = computed(() => {
+    if (!lease.value || lease.value.status !== "opened") {
+      return formatUsd(0);
+    }
+
+    const ticker = lease.value.amount.ticker;
+    const unitAssetInfo = configStore.currenciesData[`${ticker}@${lease.value.protocol}`];
+    const lpn = getLpnByProtocol(lease.value.protocol);
+
+    const unitAsset = new Dec(lease.value.amount.amount, Number(unitAssetInfo?.decimal_digits ?? 0));
+    const stableAsset = new Dec(lease.value.debt.principal, Number(lpn?.decimal_digits ?? 0)).sub(
+      new Dec(amount.value.length > 0 ? amount.value : 0)
+    );
+
+    const positionType = configStore.getPositionType(lease.value.protocol);
+    let liquidationVal: Dec;
+    if (positionType === "Long") {
+      liquidationVal = LeaseMath.calculateLiquidation(stableAsset, unitAsset);
+    } else {
+      liquidationVal = LeaseMath.calculateLiquidationShort(unitAsset, stableAsset);
+    }
+
+    return formatNumber(liquidationVal.toString(), lpn?.decimal_digits ?? 6, NATIVE_CURRENCY.symbol);
+  });
+
+  function closeDialog() {
+    const path =
+      route.matched[2].path === `/${RouteNames.LEASES}`
+        ? `/${RouteNames.LEASES}`
+        : `/${RouteNames.LEASES}/${route.params.id}`;
+    router.push(path);
+  }
+
+  return {
+    dialog,
+    assets,
+    currency,
+    isLoading,
+    loading,
+    disabled,
+    amount,
+    amountErrorMsg,
+    sliderValue,
+    calculatedBalance,
+    detbPartial,
+    liquidation,
+    debtData,
+    onSetAmount,
+    onSelectFullDebt,
+    onSelectZero,
+    handleAmountChange,
+    onSendClick,
+    closeDialog
+  };
+}
